requirements.txt
websockets>=11.0
pyaudio>=0.2.13
python-dotenv>=1.0.0
flask>=2.3.0


